requirements.txt
flask
wtforms
flask-wtf
pymongo==3.12.0
faiss-cpu
langchain-community
langchain-huggingface
sentence-transformers
langchain_mistralai
requests
google-generativeai
Pillow
cloudinary
tf-keras
